Added relative serato timecode support
Relative timecode is where a DVS such as xwax only responds to the speed
and direction the record is spinning and not the position of the needle.
This means that you can no longer skip the record or needle drop to a
new part of the song, this may be useful if you are scratching!

If you specify more that one timecode with the -t flag then can switch
between the specified timecodes by pressing Ctrl-F3 for deck A, Ctrl-F7
for deck B or Ctrl-F11 for deck C

e.g. `$ xwax -t serato_2b -t serato_rel` etc